Updated example app
