--- permission_handler/example/lib/plugin_example/permission_list.dart
@@ -1,36 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-import 'package:permission_handler/permission_handler.dart';
-
-import 'permission_widget.dart';
-
-/// Constructs a [ListView] containing [PermissionWidget] for each available
-/// permission.
-class PermissionList extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: ListView(
-          children: Permission.values
-              .where((permission) {
-                if (Platform.isIOS) {
-                  return permission != Permission.unknown &&
-                      permission != Permission.sms &&
-                      permission != Permission.storage &&
-                      permission != Permission.ignoreBatteryOptimizations &&
-                      permission != Permission.accessMediaLocation &&
-                      permission != Permission.activityRecognition &&
-                      permission != Permission.manageExternalStorage;
-                } else {
-                  return permission != Permission.unknown &&
-                      permission != Permission.mediaLibrary &&
-                      permission != Permission.photos &&
-                      permission != Permission.reminders;
-                }
-              })
-              .map((permission) => PermissionWidget(permission))
-              .toList()),
-    );
-  }
-}
