+++ permission_handler/example/pubspec.yaml
@@ -5,6 +5,10 @@ environment:
   sdk: ">=2.12.0-259.9.beta <3.0.0"
 
 dependencies:
+  baseflow_plugin_template:
+    git:
+      url: git://github.com/Baseflow/baseflow_plugin_template.git
+      ref: v2.0.0-nullsafety
   flutter:
     sdk: flutter
 
@@ -23,4 +27,6 @@ flutter:
   assets:
     - res/images/baseflow_logo_def_light-02.png
     - res/images/poweredByBaseflowLogoLight@3x.png
+    - packages/baseflow_plugin_template/logo.png
+    - packages/baseflow_plugin_template/poweredByBaseflow.png
