+++ permission_handler/pubspec.yaml
@@ -1,6 +1,6 @@
 name: permission_handler
 description: Permission plugin for Flutter. This plugin provides a cross-platform (iOS, Android) API to request and check permissions.
-version: 7.1.0
+version: 7.1.1
 homepage: https://github.com/baseflowit/flutter-permission-handler
 
 flutter:
